+++ crates/rtmp/src/chunk/mod.rs
@@ -7,6 +7,3 @@ pub use self::decoder::ChunkDecoder;
 pub use self::define::{CHUNK_SIZE, Chunk, DefinedChunkStreamID};
 pub use self::encoder::ChunkEncoder;
 pub use self::errors::{ChunkDecodeError, ChunkEncodeError};
-
-#[cfg(test)]
-mod tests;
