--- crates/rtmp/src/chunk/tests/mod.rs
@@ -1,2 +1,0 @@
-mod decoder;
-mod encoder;
