+++ crates/rtmp/src/handshake/mod.rs
@@ -2,11 +2,7 @@ mod define;
 mod digest;
 mod errors;
 mod server;
-mod utils;
 
 pub use self::define::{RTMP_HANDSHAKE_SIZE, ServerHandshakeState};
 pub use self::errors::*;
 pub use self::server::HandshakeServer;
-
-#[cfg(test)]
-mod tests;
