--- crates/rtmp/src/handshake/utils.rs
@@ -1,9 +1,0 @@
-use std::time::SystemTime;
-
-pub fn current_time() -> u32 {
-    let duration = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH);
-    match duration {
-        Ok(result) => result.as_nanos() as u32,
-        _ => 0,
-    }
-}
