+++ crates/rtmp/src/messages/mod.rs
@@ -5,6 +5,3 @@ mod parser;
 pub use self::define::{MessageTypeID, RtmpMessageData};
 pub use self::errors::MessageError;
 pub use self::parser::MessageParser;
-
-#[cfg(test)]
-mod tests;
