+++ crates/rtmp/src/netconnection/mod.rs
@@ -3,6 +3,3 @@ mod writer;
 
 pub use self::errors::NetConnectionError;
 pub use self::writer::NetConnection;
-
-#[cfg(test)]
-mod tests;
