+++ crates/rtmp/src/netstream/mod.rs
@@ -3,6 +3,3 @@ mod writer;
 
 pub use self::errors::NetStreamError;
 pub use self::writer::NetStreamWriter;
-
-#[cfg(test)]
-mod tests;
