+++ crates/rtmp/src/protocol_control_messages/mod.rs
@@ -5,6 +5,3 @@ mod writer;
 pub use self::errors::ProtocolControlMessageError;
 pub use self::reader::ProtocolControlMessageReader;
 pub use self::writer::ProtocolControlMessagesWriter;
-
-#[cfg(test)]
-mod tests;
