+++ crates/rtmp/src/session/mod.rs
@@ -1,9 +1,8 @@
 mod define;
 mod errors;
+mod handler;
 mod server_session;
 
 pub use self::errors::SessionError;
+pub use self::handler::{SessionData, SessionHandler};
 pub use self::server_session::Session;
-
-#[cfg(test)]
-mod tests;
