+++ crates/rtmp/src/user_control_messages/mod.rs
@@ -4,6 +4,3 @@ mod writer;
 
 pub use self::errors::EventMessagesError;
 pub use self::writer::EventMessagesWriter;
-
-#[cfg(test)]
-mod tests;
