+++ crates/rtmp/src/user_control_messages/writer.rs
@@ -26,3 +26,28 @@ impl EventMessagesWriter {
         Ok(())
     }
 }
+
+#[cfg(test)]
+#[cfg_attr(all(test, coverage_nightly), coverage(off))]
+mod tests {
+    use bytes::{BufMut, Bytes, BytesMut};
+
+    use super::*;
+    use crate::chunk::ChunkDecoder;
+
+    #[test]
+    fn test_write_stream_begin() {
+        let mut buf = BytesMut::new();
+        let encoder = ChunkEncoder::default();
+
+        EventMessagesWriter::write_stream_begin(&encoder, &mut (&mut buf).writer(), 1).unwrap();
+
+        let mut decoder = ChunkDecoder::default();
+
+        let chunk = decoder.read_chunk(&mut buf).expect("read chunk").expect("chunk");
+        assert_eq!(chunk.basic_header.chunk_stream_id, 0x02);
+        assert_eq!(chunk.message_header.msg_type_id as u8, 0x04);
+        assert_eq!(chunk.message_header.msg_stream_id, 0);
+        assert_eq!(chunk.payload, Bytes::from(vec![0x00, 0x00, 0x00, 0x00, 0x00, 0x01]));
+    }
+}
